% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubeplay.R
\name{tubeplay}
\alias{tubeplay}
\title{Play YouTube moview at RStudioServer}
\usage{
tubeplay(url = "https://www.youtube.com/watch?v=iOFZKwv_LfA", width = "560",
  height = "315", background = "white", viewer = getOption("viewer",
  utils::browseURL))
}
\arguments{
\item{url}{url that you want show Toutube movie}

\item{width}{width of movie}

\item{height}{height of movie}

\item{background}{background color}

\item{viewer}{set viewer. See ?htmltools::html_print .}
}
\description{
Play YouTube moview at RStudioServer
}
\examples{
tubeplay() # play default set movie
tubeplay("https://www.youtube.com/watch?v=0E00Zuayv9Q") # play "PPAP"
}

