% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubeplayR-package.R
\docType{package}
\name{tubeplayR-package}
\alias{tubeplayR}
\alias{tubeplayR-package}
\title{tubeplayR: play youtube by R}
\description{
play youtube on RStudio.
}
\author{
\strong{Maintainer}: Kazuhiro Maeda \email{kazu@kz-md.net}

}
\keyword{internal}
