% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubeplay.R
\name{tubeplay}
\alias{tubeplay}
\title{Play YouTube moview at RStudioServer}
\usage{
tubeplay(url = "https://www.youtube.com/watch?v=iOFZKwv_LfA",
  autoplay = TRUE, width = "100\%", height = "100\%",
  background = "white", viewer = getOption("viewer", utils::browseURL))
}
\arguments{
\item{url}{url that you want show Toutube movie}

\item{autoplay}{Logical. Whether the movie play auto.}

\item{width}{ignored.}

\item{height}{ignored.}

\item{background}{ignored.}

\item{viewer}{set viewer. See ?htmltools::html_print .}
}
\description{
Play YouTube moview at RStudioServer
}
\examples{
tubeplay() # play default set movie
tubeplay("https://www.youtube.com/watch?v=0E00Zuayv9Q") # play "PPAP"
}

