% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubeplay.R
\name{tubeplay}
\alias{tubeplay}
\title{Play YouTube movie at RStudio}
\usage{
tubeplay(
  url = "https://www.youtube.com/watch?v=XSLhsjepelI",
  viewer = getOption("viewer", utils::browseURL),
  add_history = TRUE
)
}
\arguments{
\item{url}{url that you want show YouTube movie}

\item{viewer}{set viewer. See ?htmltools::html_print.}

\item{add_history}{whether add to play history or not.}
}
\description{
Play YouTube movie at RStudio
}
\examples{
tubeplay() # play default set movie
tubeplay("https://www.youtube.com/watch?v=0E00Zuayv9Q") # play "PPAP"
}
